+++ pkg/application/app.go
@@ -3,6 +3,7 @@ package application
 import (
 	"bytes"
 	"fmt"
+	"os"
 	"regexp"
 	"strings"
 
@@ -14,6 +15,10 @@ import (
 	"sigs.k8s.io/e2e-framework/klient/decoder"
 
 	"github.com/giantswarm/clustertest/pkg/organization"
+)
+
+const (
+	defaultValuesContents = "\n"
 )
 
 // If commit SHA based version we'll change the catalog
@@ -51,7 +56,7 @@ func New(installName string, appName string) *Application {
 		ClusterName:  "",
 		Version:      "",
 		Catalog:      "cluster",
-		Values:       "\n",
+		Values:       defaultValuesContents,
 		InCluster:    true,
 		Organization: *organization.New("giantswarm"),
 	}
@@ -112,6 +117,17 @@ func (a *Application) MustWithValues(values string, config *TemplateValues) *App
 //
 // The file supports templating using Go template strings and uses values provided in `config` to replace placeholders.
 func (a *Application) WithValuesFile(filePath string, config *TemplateValues) (*Application, error) {
+	// We need to check that the values file actually has contents otherwise kubectl-gs fails to build the Application
+	fileBytes, err := os.ReadFile(filePath)
+	if err != nil && !os.IsNotExist(err) {
+		return nil, err
+	}
+	if len(fileBytes) == 0 {
+		// Empty file so we'll set it to the default contents
+		a.Values = defaultValuesContents
+		return a, nil
+	}
+
 	values, err := parseTemplateFile(filePath, config)
 	if err != nil {
 		return nil, err