+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Safely handle an empty values file being provided by replacing it with a single newline character
+
 ## [0.14.0] - 2023-12-04
 
 ### Changed